Renamed redis_utils module
